requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask==3.0.0
python-dotenv==1.0.0
pymongo[srv]
Flask-PyMongo